# a9_openapi.yaml
openapi: 3.0.3
info:
  title: Twira
  version: "1.0"
...
paths:
  /account/login:
    get:
      operationId: loginPage
      summary: 'Login page'
      tags:
        - account
      responses:
        '200':
          description: 'Login page'
    post:
      operationId: login
      summary: 'Logs user into the system'
      tags:
        - account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/LoginRequest'
      responses:
        '200':
          description: Successful login
        '403':
          description: Invalid email or password
...
  /profile:
    get:
      operationId: selfProfilePage
      summary: Shows the logged in user's profile page
      tags:
        - profile
      responses:
        '200':
          description: Profile page
        '401':
          description: Not logged in
    patch:
      operationId: editProfile
      summary: Edits a user's profile
      tags:
        - profile
      requestBody:
        description: Profile information
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditProfile'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EditProfile'
      responses:
        '200':
          description: Operation successful
        '401':
          description: Not logged in
...
  /post/create:
    post:
      operationId: createPost
      summary: Creates a post
      tags:
        - post
      requestBody:
        description: Post contents
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreatePost'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/CreatePost'
      responses:
        '200':
          description: Operation successful
        '401':
          description: Not logged in
    get:
      operationId: createPostPage
      summary: Shows the create post page
      tags:
        - post
      responses:
        '200':
          description: Post creation page
        '401':
          description: Not logged in
  /post/{id}:
    get:
      operationId: viewPostPage
      summary: Shows the specified post
      tags:
        - post
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the post
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Post page
        '403':
          description: Not allowed to view this post
        '404':
          description: Post not found
    patch:
      operationId: editPost
      summary: Edits the specified post
      tags:
        - post
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the post
          schema:
            type: integer
            format: int64
      requestBody:
        description: Post edit
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EditPost'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/EditPost'
      responses:
        '200':
          description: Operation successful
        '403':
          description: Not allowed to edit this post
        '404':
          description: Post not found
    delete:
      operationId: deletePost
      summary: Deletes the specified post
      tags:
        - post
      parameters:
        - name: id
          in: path
          required: true
          description: The ID of the post
          schema:
            type: integer
            format: int64
      responses:
        '200':
          description: Operation successful
        '403':
          description: Not allowed to delete this post
        '404':
          description: Post not found
...
components:
  schemas:
...
    LoginRequest:
      type: object
      properties:
        email:
          type: string
          example: example@email.com
        password:
          type: string
          example: hunter2
...
    CreatePost:
      type: object
      properties:
        text:
          type: string
          description: The text of the post (optional if it's a video or image post)
        contentUrl:
          type: string
          format: url
          description: The URL of the media content (must not be included on text posts)
    EditPost:
      type: object
      properties:
        text:
          type: string
          description: The text of the post (optional if it's a video or image post)
...
    EditProfile:
      type: object
      properties:
        birthDate:
          type: string
          format: date
          example: '1970-01-01'
        phone:
          type: string
          example: '12345'
        bio:
          type: string
          example: 'This is some information about the user'
        publicProfile:
          type: boolean
          example: true